requirements-dev.txt
-r requirements.txt

# Development and testing dependencies
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.2.7
black==23.7.0
flake8==6.0.0
isort==5.12.0